hot-fix: restructure the project files
