+++ src/pages/Auth/Login.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { Link } from "react-router-dom";
-import useLoadingStore from "../store/loadingstore";
+import useLoadingStore from "../../store/loadingstore";
 import { useNavigate } from "react-router";
 const Login = () => {
   const [email, setEmail] = useState("");
