+++ src/pages/Home/Home.jsx
@@ -1,6 +1,6 @@
-import Hero from "./Hero";
-import Features from "./Features";
-import HowItWorks from "./HowItWorks";
+import Hero from "../../components/Hero";
+import Features from "../../components/Features";
+import HowItWorks from "../../components/HowItWorks";
 import Header from "../../components/Header";
 import Footer from "../../components/Footer";
 const Home = () => {
